Weather, first pass
git-svn-id: svn+ssh://50.28.74.132/home/luminari/svn/luminari/myproject/trunk@3576 dda7f814-ccd3-4591-a6d7-1bd23e5e64af

+++ perlin.h
@@ -28,8 +28,8 @@ void normalize3(double *);
 void normalize2(double *);
 
 double PerlinNoise1D(int idx, double,double,double,int);
-double PerlinNoise2D(int idx, double,double,double,double,double,int);
-double PerlinNoise3D(int idx, double,double,double,double,double,double,int);
+double PerlinNoise2D(int idx, double,double,double,double,int,int);
+double PerlinNoise3D(int idx, double,double,double,double,double,int,int);
 double RidgedMultifractal2D( int idx, double x, double y, double H, double lacunarity,
                              double octaves, double offset, double gain );
